googlegemini/requirements.txt
google-genai
pillow
pydantic